+++ packages/app/src/pages/over-risiconiveaus.tsx
@@ -1,107 +1,55 @@
-import Head from 'next/head';
-import { Box } from '~/components/base';
-import { RichContent } from '~/components/cms/rich-content';
-import { CollapsibleSection } from '~/components/collapsible';
-import { Heading } from '~/components/typography';
-import { Content } from '~/domain/layout/content';
+import { vrData } from '~/data/vr';
 import { Layout } from '~/domain/layout/layout';
+import { NationalLayout } from '~/domain/layout/national-layout';
 import { useIntl } from '~/intl';
 import {
   createGetStaticProps,
   StaticProps,
 } from '~/static-props/create-get-static-props';
 import {
-  createGetContent,
   getLastGeneratedDate,
+  loadAndSortVrData,
+  selectCustomData,
+  selectNlPageMetricData,
 } from '~/static-props/get-data';
-import { CollapsibleList, RichContentBlock } from '~/types/cms';
-import { getSkipLinkId } from '~/utils/skip-links';
-
-interface OverRisiconiveausData {
-  title: string | null;
-  description: RichContentBlock[] | null;
-  collapsibleList: CollapsibleList[];
-}
 
 export const getStaticProps = createGetStaticProps(
   getLastGeneratedDate,
-  createGetContent<OverRisiconiveausData>((_context) => {
-    //@TODO We need to switch this from process.env to context as soon as we use i18n routing
-    // const { locale } = context;
-    const locale = process.env.NEXT_PUBLIC_LOCALE;
-    return `*[_type == 'overRisicoNiveaus']{
-      ...,
-      "description": {
-        "_type": description._type,
-        "${locale}": [
-          ...description.${locale}[]
-          {
-            ...,
-            "asset": asset->
-           },
-        ]
-      },
-      "collapsibleList": [...collapsibleList[]
-        {
-          ...,
-          "content": {
-            ...content,
-            "${locale}": [
-              ...content.${locale}[]
-              {
-                ...,
-                "asset": asset->
-               },
-            ]
-          }
-      }]
-    }[0]
-    `;
+  selectNlPageMetricData(),
+  selectCustomData(() => {
+    const risiconiveaus = vrData.map((vr) => {
+      const vrData = loadAndSortVrData(vr.code);
+      return {
+        data: vrData.escalation_level,
+        safetyRegionName: vr.name,
+        vrCode: vr.code,
+      };
+    });
+
+    return {
+      risiconiveaus,
+    };
   })
 );
 
 const OverRisicoNiveaus = (props: StaticProps<typeof getStaticProps>) => {
   const { siteText } = useIntl();
-  const { content, lastGenerated } = props;
+  const { selectedNlData: data, lastGenerated, risiconiveaus } = props;
+
+  const text = siteText.rioolwater_metingen;
+
+  const metadata = {
+    ...siteText.nationaal_metadata,
+    title: text.metadata.title,
+    description: text.metadata.description,
+  };
 
   return (
-    <Layout
-      {...siteText.over_risiconiveaus_metadata}
-      lastGenerated={lastGenerated}
-    >
-      <Head>
-        <link
-          key="dc-type"
-          rel="dcterms:type"
-          href="https://standaarden.overheid.nl/owms/terms/webpagina"
-        />
-        <link
-          key="dc-type-title"
-          rel="dcterms:type"
-          href="https://standaarden.overheid.nl/owms/terms/webpagina"
-          title="webpagina"
-        />
-      </Head>
-      <Content>
-        {content.title && <Heading level={1}>{content.title}</Heading>}
-        {content.description && <RichContent blocks={content.description} />}
-        {content.collapsibleList && (
-          <article>
-            {content.collapsibleList.map((item) => {
-              const id = getSkipLinkId(item.title);
-              return item.content ? (
-                <CollapsibleSection key={id} id={id} summary={item.title}>
-                  {item.content && (
-                    <Box mt={3}>
-                      <RichContent blocks={item.content} />
-                    </Box>
-                  )}
-                </CollapsibleSection>
-              ) : null;
-            })}
-          </article>
-        )}
-      </Content>
+    <Layout {...metadata} lastGenerated={lastGenerated}>
+      <NationalLayout
+        data={data}
+        lastGenerated={lastGenerated}
+      ></NationalLayout>
     </Layout>
   );
 };
